@@ -1,9 +1,8 @@
 ﻿using Frost.Net.Interoperability;
 
-namespace Frost.Net.Synchronization.Interface
+namespace Frost.Net.Synchronization.Interface;
+
+internal interface ISynchronizableEvent : ISynchronizable
 {
-    internal interface ISynchronizableEvent : ISynchronizable
-    {
-        public void Reset() => FrostApi.Synchronizable.ResetEvent(Handle);
-    }
+    public void Reset() => FrostApi.Synchronizable.ResetEvent(Handle);
 }
