@@ -1,8 +1,9 @@
-﻿namespace Frost.Net;
-
-using Frost.Net.Interoperability;
+﻿using Frost.Net.Interoperability;
 using Frost.Net.Models;
 using Frost.Net.Utilities;
+using System.Reflection;
+
+namespace Frost.Net;
 
 public static class EventSystem
 {
@@ -11,12 +12,45 @@ public static class EventSystem
 
 
 
-	static EventSystem()
+	/// <summary>
+	/// Subscribe method via reflections.
+	/// </summary>
+	/// <param name="activationLayers">Event layers that activate the method.</param>
+	/// <param name="method">Method info of target method.</param>
+	/// <param name="obj">Optional this parameter of the method.</param>
+	public static void Subscribe(Layers activationLayers, MethodInfo method, object? obj = null)
 	{
-		_logEventsTag = FrostApi.Logging.GetLogEventTag();
-		FrostApi.EventSystem.SubscribeRelay(InteropRelay);
+		var type = method.GetParameters().First().ParameterType.GetElementType();
+		var delegateType = typeof(Handler<>).MakeGenericType(type);
+		var hash = (ulong)type.GetHashCode();
+
+		// Register tagged relay
+		typeof(EventSystem)
+			?.GetMethod(nameof(RegisterTaggedRelay))
+			?.MakeGenericMethod(method.GetParameters().First().ParameterType).Invoke(null, new object[] { hash });
+
+		// Find or create tagged collection
+		if (!_handlers.TryGetValue(hash, out var collection))
+		{
+			collection = new Dictionary<ulong, object>();
+			_handlers[hash] = collection;
+		}
+		// typeof(Handler<>).MakeGenericType(type)
+		// Find or create layer collection
+		if (collection.TryGetValue(activationLayers, out var handlerDelegate))
+			collection[activationLayers] = Delegate.Combine(handlerDelegate as Delegate, method.CreateDelegate(delegateType));
+		else
+			collection[activationLayers] = method.CreateDelegate(delegateType);
+		
 	}
 
+	/// <summary>
+	/// Subscribe method via reflections.
+	/// </summary>
+	/// <typeparam name="T">Type of event data./typeparam>
+	/// <param name="tag">Type tag, by default (ulong)typeof(T).GetHashCode()</param>
+	/// <param name="activationLayers">Event layers that activate the method.</param>
+	/// <param name="handler">Subscribed delegate.</param>
 	public static void Subscribe<T>(ulong tag, Layers activationLayers, Handler<T> handler) where T : struct
 	{
 		// Register tagged relay
@@ -35,10 +69,24 @@ public static class EventSystem
 		else
 			collection[activationLayers] = handler;
 	}
+
+	/// <summary>
+	/// Subscribe method via reflections.
+	/// </summary>
+	/// <typeparam name="T">Type of event data./typeparam>
+	/// <param name="activationLayers">Event layers that activate the method.</param>
+	/// <param name="handler">Subscribed delegate.</param>
 	public static void Subscribe<T>(Layers activationLayers, Handler<T> handler) where T : struct =>
 		Subscribe((ulong)typeof(T).GetHashCode(), activationLayers, handler);
 
 
+
+	/// <summary>
+	/// Emits an event on given layers.
+	/// </summary>
+	/// <typeparam name="T">Type of event data.</typeparam>
+	/// <param name="activationLayers">Layers on which event is emitted on</param>
+	/// <param name="eventData">Event data given to handlers</param>
 	public static void Emit<T>(Layers activationLayers, ref T eventData) where T : struct
 	{
 		unsafe
@@ -49,8 +97,18 @@ public static class EventSystem
 #pragma warning restore CS8500
 		}
 	}
+
+	/// <summary>
+	/// Emits an event on given layers.
+	/// </summary>
+	/// <typeparam name="T">Type of event data.</typeparam>
+	/// <param name="activationLayers">Layers on which event is emitted on</param>
+	/// <param name="eventData">Event data given to handlers</param>
 	public static void Emit<T>(Layers activationLayers, T eventData) where T : struct =>
 		Emit(activationLayers, ref eventData);
+
+
+
 	private static void EmitInternal<T>(Layers activationLayers, ref T e) where T : struct
 	{
 		if (!_handlers.TryGetValue((ulong)typeof(T).GetHashCode(), out var collection))
@@ -62,7 +120,7 @@ public static class EventSystem
 		{
 			if ((pair.Key & activationLayers.Value) == 0)
 				continue;
-
+			var x = (pair.Value as Handler<T>);
 			(pair.Value as Handler<T>)?.Invoke(ref e);
 		}
 	}
@@ -139,7 +197,11 @@ public static class EventSystem
 	private static void RegisterTaggedRelay<T>(ulong tag) where T : struct =>
 		_relays[tag] = TaggedDirectRelay<T>;
 
-
+	static EventSystem()
+	{
+		_logEventsTag = FrostApi.Logging.GetLogEventTag();
+		FrostApi.EventSystem.SubscribeRelay(InteropRelay);
+	}
 
 	private delegate void Relay(ulong tag, ulong activationLayers, IntPtr pData);
 
